+++ changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased - 2019-01-01
+## Unreleased - 2020-01-01
 - [Commits](https://github.com/unity-game-framework/ugf-utf8json/compare/0.0.0...0.0.0)
 - [Milestone](https://github.com/unity-game-framework/ugf-utf8json/milestone/0?closed=1)
 
@@ -25,6 +25,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 - Nothing.
+
+## 4.1.0-preview - 2020-01-12
+- [Commits](https://github.com/unity-game-framework/ugf-utf8json/compare/4.0.0-preview...4.1.0-preview)
+- [Milestone](https://github.com/unity-game-framework/ugf-utf8json/milestone/15?closed=1)
+
+### Added
+- `Utf8JsonBuiltinResolverAsset` and `Utf8JsonUnityResolverAsset`.
+
+### Deprecated
+- `Utf8JsonUtility.CreateDefaultResolver` has been deprecated.
 
 ## 4.0.0-preview - 2020-01-11
 - [Commits](https://github.com/unity-game-framework/ugf-utf8json/compare/3.2.2-preview...4.0.0-preview)